[GM3Restyling] Use `ToolbarMenuButton` in Console panel.
Prior to this CL we had three occurences of `ToolbarButton`s that would
turn into selects with a menu attached to it, that would not use the
`ToolbarMenuButton`, but rather use a plain `ToolbarButton` with custom
menu logic:

1. The level selector in the Console panel.
2. The experimental request type selector in the Network panel.
3. The experimental 'More filters' selector in the Network panel.

With this CL we update the level selector to use the proper
`ToolbarMenuButton` class to enforce the invariant that all of
these select buttons in toolbars now consistently use the same
implementation (and the same menu logic and appearance). We
cannot change the two buttons in the Network panel right now,
since they rely on the `keepOpen` feature that only the soft
context menu provides.

Bug: 330213829
Change-Id: I8c4bb26b284c48e4370cb340bb6bcca3a3df9e7a
Screenshot: https://i.imgur.com/1vVO3dt.png
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5481689
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>