RPP: remove old engine auto-deprecating line numbers
For a subset of events, the old TimelineModel would automatically
subtract 1 from the line numbers to make them 0 indexed. We don't want
to rely on this behaviour (which we implicitly do right now as we run
both engines).

This CL removes that behaviour from the legacy model and updates
TimelineUIUtils with a helper that can read line numbers from events. It
will 0-index line numbers for the events that we know are 1 indexed.

Bug: 336734268
Change-Id: Ibaca450e845fe6b1b65db2fa875b72415308ad2b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5488980
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>