RPP: helper for getting an event's URL
We rely on EventOnTimelineData to store URLs for events, but this is
going to be removed as we migrate to the new engine.

This CL introduces a helper which can get the URL from an event for any
given event from the new engine.

A follow-up CL will use this helper, and remove EventOnTimelineData.url

Bug: 336739509
Change-Id: I4dccd93833dfbc23b20c418ef1ebfefea066a7db
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5490523
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>