RPP: add experiment to run old engine
This will allow us to locally disable the old engine and check for
breakages, without the fear of breaking users. Once we are confident we
can make the change permanent and remove the old model.

Note that I named the experiment "Enable", rather than "Disable", to
avoid a confusing double negative. But the experiment is enabled by
default, so the default experience for users here is to run the old
engine.

Bug: 40287735
Change-Id: I2eeefc1d3815f7a0dfcb43ba1eb3cf460c3176af
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5497335
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>