RPP: update TimelineUIUtils to take only new events
This CL updates the type def of the main method in TimelineUIUtils that
builds details for events. It now only takes new events, and also
insists on trace data being present.

Within that function there are a lot of untyped references to
`event.args` and `event.args.data`. I want to solve that and make it
type safe, but for now I have cast the objects to allow me to not have
to update all these references at once (there are 50+). This is not a
regression as it has no less type safety than prior to this CL.

Bug: 40287735
Change-Id: I652d1b9adf04170b6ccdac9b4772ea76ae9899b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5505713
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>