RPP: store ProfileId & SampleIndex with ProfileCall
As part of my exploration into how we can reliably serialize data that
includes the ability to store data attached to events (think
annotations), I have been exploring how to ensure that a ProfileCall
event can be linked back to a single trace event in the trace.

For a given ProfileCall we can construct a unique reference to it if we
combine:

1. The ProfileId
2. The ProcessId (because the same ProfileId could be reused across
   processes, so isn't unique on its own)
3. The index of the sample that created this call
4. The ID of the node that references the function call of this
   ProfileCall

This CL threads through the necessary data in order to create
ProfileCall events with this data to be used in the future.

Bug: 339163152
Change-Id: I3bfebaf891a1f9a30457e2aa687559d2808f999f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5518718
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>