[sources] Support saving files with binary content
This file utilizes the new "isBase64" flag for the "save" host binding
introduced with https://crrev.com/c/5522321.

Instead of decoding base64-encoded in the frontend, we'll let the
backend handle this correctly.

As a follow-up we can remove the "Save Image" variant (which creates
an <a>-nchor and clicks it). We'll also be able to support Wasm
modules, but they need special handling.

Bug: 40784130
Change-Id: I2dd762f9c7730c8ad30a7f61aa8f04b9f57b6673
Fixed: 337191314
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5534700
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>