[autofill in devtools] Add german test adress
This address does not follow the new german model. This is because the
model is not rolled out yet. However Germany already does not have state
information, which is a difference when compared to the US for example.

This means that for the purpose of this feature, it still makes sense
(albeit less) to have Germany as its own entry. It is also important
to have it now so that we can offer a large list of addresses to
developers.

Follow up CL:

1. Add a german address for the new model.
2. Update chrome to branch between the two addresses depending
   on whether the new model is available.

Bug: b/40270486
Change-Id: I1c841a7be3df37baeb151e74cb268ea28fb8115c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5534706
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Norge Vizcay <vizcay@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>