Fix the miscalculation in request timing when HAR is imported
This CL fixes the bug which miscalculates request timestamps displayed
at the timing view in the network panel.

When the HAR is exported, the `blocked` field includes the
`blocked_queueing` time. When the HAR is imported, the current codebase
mistakenly put the `blocked` field to `lastEntly` despite
`blocked_queueing` time is taken into account in the `requestTime`, and
it results in other subsequent timing fields are miscalculated.

This is a blocker to support service workers in HAR.

Bug: 340109211
Change-Id: I321f4d3b10a8c26b479502a8f7da48551ab8465f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5534345
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>