Only use pause location for breakpoint if it is valid
Frequently when paused on DOM breakpoints or CSP breakpoints
the debugger stops in a location where it is not valid to set
a breakpoint. When selecting "Never Pause Here" at such locations,
my previous change caused a breakpoint to be set on the following
line. With this change, the debugger checks for this and falls
back to setting it at the default location for the line.

Bug: b:40924349
Change-Id: I8a62a0f5b4ec6573e896a024379253d92d8917e0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5541794
Reviewed-by: Simon Zünd <szuend@chromium.org>
Auto-Submit: Eric Leese <leese@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>