[RenderBlocking] Only consider body requests that are high priority
The render blocking audit in Lighthouse actually has a second level of
filtering that only includes requests that are in the FCP simulation
graph. The FCP simulation graph filters out low priority requests:
http://shortn/_rUloOEiNu9

For requests marked `in_body_parser_blocking` we should ensure that
the priority is high enough to be considered render blocking. There are
exceptions to the logic added but this is the definition Lighthouse has
been using so this is at least a good place to start.

This CL changes nothing about the behavior of requests marked `blocking`

Bug: 328489888
Change-Id: I7700118e78f9f36f62ba5acfb2da68928ccc1879
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5537846
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>