Add service worker related information to the custom fields in HAR
In a current HAR format, any information about service worker are not
set. This was repeatedly reported by web developers and makes their
debugging difficult.

For HAR, spec hasn't touched for a long time and it's not actively
maintained. So it makes sense to add service worker information to
custom fields, which are prefixed with underscore.

After this CL, the HAR file can have following fields.
- _fetchedViaServiceWorker
- _responseCacheStorageCacheName
- _serviceWorkerResponseSource

After this CL, the initiator column in the network panel shows "Service
Worker" is the resource was fetched via service worker. But this doesn't
mean all of service worker info are available to see in the network
panel when importing a HAR file. `_responseCacheStorageCacheName` and
`_serviceWorkerResponseSource` are not displayed yet, those will be
supported after adding service worker timestamps in request.timing.

The static routing API information is out of scope in this CL.

Bug: 40286129
Change-Id: I835ab4ad0db5817fc00d0596e5d23c56ab7eb985
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5531304
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>