[sources] Support 'Save as' context menu for WASM files
WASM UISoureCodes are somewhat special in that they return a
DisassemblyInfo in their `requestContent` implementation but not the
actual raw bytes.

This CL adds special handling by checking if a UISourceCode has
an underlying WASM script, requests it from the backend, encodes
it as Base64 and passes it along to the 'save' host binding.

This is somewhat convoluted but it's not a much used workflow.

R=pfaffe@chromium.org

Fixed: 40784130
Change-Id: I5cdfc511489cb8778dd933f80cdb4619df26c2c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5534705
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>