Fix streaming response content view for requests that just started
This CL fixes a bug where we would tell the user that a request
"Has no response data available" even though it has. This happens
when we a user opens a request that has just started but no
"responseReceived" event with the initial headers has arrived in
DevTools yet.

In that case we would create a StreamingContentData with mime type
'application/octet-stream', which then later would result in no
"Response" tab since we don't show those for binary content.

The fix is to delay creation of the "StreamingContentData" until
we have at least received the initial HTTP headers of the response
so we know MIME and charset.

R=bmeurer@chromium.org

Fixed: 338340752
Change-Id: Ifd422b87c7ec398509732fec02cda8235d7ddf2c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5545599
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>