Use issueTime when calculating load times for prefetched documents
When navigating to a prefetched page, requestStart is much earlier
than when the actual navigation starts, leading to the durations
displayed in the Network summary tab (like DOMContentLoaded) being
much longer than it actually is. We use issueTime instead, which
corresponds more closely to when the navigation starts.

Bug: 339498841
Change-Id: I7cd25a5ac3ae9909836c0bf995fd514e14f2d83f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5545675
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>