[browser_protocol] Remove dependency from roll script
It got removed from Chromium and the roll script fails now because of it. It also failed without this dependency in the past when we had to add it. We will probably need to pick it from somewhere else if it is still needed.

Bug: 340813542
Change-Id: If33364a9ec34220afb210ea02c02e683930cb0b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5545190
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>