ExtensionTraceDataHandler ignores non-obj perf marks
This code accidentally relied on the details being something that is
`JSON.parse`d as an object, so we can then use the 'in' operator to
search for it.

The fix in this context is to wrap the code in a try catch and return
null if it errors - which means either the details data was not valid
JSON, or it wasn't of the structure we need for the sake of extensions.
Both of these are OK and we just need the extension code to ignore the
event in that case.

Fixed: 341254293
Change-Id: Ic81c006d039b9a240e2c8d7ce8d37d0985fe022b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5548337
Reviewed-by: Victor Porof <victorporof@chromium.org>
Commit-Queue: Victor Porof <victorporof@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
(cherry picked from commit c1d688bcb2e1d1ebcebbe07e48ddcfa90228b138)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5569110
Reviewed-by: Andres Olivares <andoli@chromium.org>