RPP: Add Totals row and Tooltip to Selector Stats
The %-of-Slow-Path-Non-Matches aka FastRejectPercentage is hard to
understand, so a detailed tooltip is added to explain.

A new (Totals) row illuminates the totals of the entire datagrid.

And the column widths were tweaked to better fit the header cell text.

Screenshot: https://imgur.com/a/qR2Geck

Bug: 324282954
Change-Id: I515caf3a7aa9e00593d31534451ae9f5a856e2aa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5560268
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Guangyue Xu <guangyue.xu@microsoft.com>
Commit-Queue: Paul Irish <paulirish@chromium.org>