[Compression Dictionary Transport] Update the error message
This CL updates the error message on the Issues tab for
UnexpectedContentDictionaryHeader error.

Before https://crrev.com/c/5552563, Chromium used this error when the
`Content-Dictionary` header was not set correctly for a dictionary
compressed response. But after that CL, Chromium uses this error when
the header of dictionary compressed stream was not set correctly.

So this CL updates the error message used in the Issues tab of DevTools.

Bug: 333756098
Change-Id: Id30e55f1f4ae1ff1ac11026c2829a8867d0b7abe
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5561846
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>