[sdk] Move manual WASM disassembly from SourceFrame.ts to Script.ts
Script#requestContent should always return a WasmDisassemblyInfo (or
an error). Who does the actual disassembling is an implementation
detail that the `SourceFrame` should not have to worry about.

We loose the progress indicator but according to @pfaffe this did
not work properly in the first place.

R=pfaffe@chromium.org

Bug: 342529015
Change-Id: I6a21548e29b189c41ac231817793bd171bc0a08e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5569109
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>