[workspace] Change 'Project' interface to produce a 'ContentData'
The two big implementers (FileSystem and ContentProviderBasedProject)
already use ContentData under the hood. Only SnippetFileSystem is
missing.

In the end this CL pushes 'asDeferredContent' conversion further
up the call-chain. But because the 'Project' interface is used
so widespread, we have to do it in one larger CL.

R=kimanh@chromium.org

Bug: 342529015
Change-Id: I8552888f7d2b465ba0069892c110f9c49039e2bd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5580039
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>