[RPP] Add live metrics experiment to the performance panel
This uses the web-vitals library to install live metric listeners on the
target page and surface them on the performance panel landing page.

Other supplemental data is included such as the LCP element.

The initial UI is extremely simple but we will continue to iterate on
that behind the experiment.

Histogram CL: https://crrev.com/c/5579296

Bug: 313906462
Change-Id: Idb6aaaa1a0d5f542a1767218e3f8f4483d010658
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5568374
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>