[cleanup] An explicit parameter to control throttler scheduling behavior.
Currently throttler schedules the task immediately if it doesn't have
pending tasks or if the second parameter is true.

This default behavior is often what we want for throttling heavy
operations but is not suitable for debouncing. It is also somewhat
surprising that setting asSoonAsPossible to false can still schedule
immediately.

This will allow to simplify VE logging implementation and test, where to
ensure debouncing we currently schedule an empty task before scheduling
the real task, so that the latter will never run immediately.

Bug: 343845140
Change-Id: If5888ed491ef3d9693f2df714e8e2c06e33e2ff1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5594154
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>