RPP: enable overlays to support frames
The initial Overlays implementation assumed they would always be
attached to trace events, but frames on the panel are `TimelineFrame`
instances.

(Long term they won't be, once the new track is powered by Animation
Frames - see crbug.com/345144583 - but that is far enough away that we
need to support the existing frames implementation).

This CL updates the Overlays code to deal with frames and enable them to
be used as the entry provided to an overlay. It also takes the
opportunity to do some tidying up and change the API slightly - rather
than the user providing the `entryChart` this is now deduced within the
Overlays code based on the type of entry provided.

Fixed: 344771436
Change-Id: I1f0ce409e1de9dd6b51a8ad3157ecd3cd8a51a9c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5601129
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>