[sdk] Decode bindings for 'Scopes' source map proposal
This CL adds the last, and arguably most complicated bit, of the
"Scopes" proposal: bindings. Bindings encode what expression to use
to produce the value of a variable in a given generated range.

Since the expression might not be the same throughout the whole range,
the encoding allows for sub-ranges. This means that a binding for a
given variable could either be a single expression or a list of
sub-ranges with each sub-range having their own expression.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I879caed0ecf8bc377b600357ffd98989650ebb34
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5598789
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>