Revert "[RPP] Move EventUICategory to trace helpers and use it to check if the parent hidden with EntriesFilter is visible"
This reverts commit 54723fa2c4ed1b9f702d17234b4b93d361f0ff71.

Reason for revert: models/trace needs to be isolated from the rest of DevTools, because it is packaged up into an npm package that Lighthouse depends on, and in the future more tools might use. This CL means that that package has to include a bunch of DevTools (i18n, ThemeSupport, Root) which we do not want. Nothing inside models/trace should import from outside of it (we do have some exceptions to this around CPU Profiling IIRC, but this needs to be fixed).

Original change's description:
> [RPP] Move EventUICategory to trace helpers and use it to check if the parent hidden with EntriesFilter is visible
>
> Moving EventUICategory out of timeline so we can use its' methods in the trace module without introducing a circular dependency.
>
> Also use entryIsVisibleInTimeline method to fix the bug where we tried to add "collapsed" arrow to entries that were not visible.
>
> Bug: 338977625
> Change-Id: I0459d9b711f4f2c5c90f9d735079c2d57e99af76
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5596375
> Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
> Reviewed-by: Andres Olivares <andoli@chromium.org>

Bug: 338977625
Change-Id: I6ac944979e6dd09ed8367c320edfce9297400ace
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5601395
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>