[trace engine] Don't throw when no SynEvents manager was created
Registering synthetic events is needed to power modification storing
in the Perf panel. However, the trace engine should remain usable
outside the Perf panel. Thus, we cannot enforce a instance of the
SyntheticEvents manager to be created from the trace engine. Instead,
when running the perf panel we initialize the manager and register in
it, otherwise we don't register any data.

Bug: 339163152
Change-Id: Ie504264d3bef34f9256df843b7973a1f9c4c9ac3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5604302
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>