[sdk] Hook up scopes proposal parsing to source map parsing
This CL wires up the decoding of "original scopes" and "generated
ranges" to the general source map parsing (but behind an experimental
flag).

As part of this CL we rename the experimental bloomberg scopes
("pasta source maps") so we can reuse the "parseScopes" name.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: I9b776784199ccf83162e62313700613219c7a471
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5612392
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>