[sdk] Use source map 'scopes' to expand inlined call frames on pause
This CL extends call frame expansion in the DebuggerModel when we pause.
We not only invoke the callback installed by the DWARF language
plugin, but also utilize the 'scopes' information from the source map.

Note that calculating the actual position is done via LiveLocation
and will happen in a follow-up CL where we take the now specified
"inlineFrameIndex" into account when translating from a
"raw location" to a UILocation.

We trigger the logic only if the experiment is enabled since the
implementation is not ideal: We'll wait for all relevant source maps
to finish loading before finishing up the paused details. Before we
can mark the experiment as stable or enable it by default we need to
make call frame expansion asynchronous so we don't block the pause UI
on it.

Generally, DevTools also supports enabling/disabling source maps
throughout a session, so making call frame expansion asynchronous
and/or event based will fix that as well. But to get something out
the door and verify the current 'scopes' proposal format, this is
good enough for now.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: I01d3e87cc0d81585a1fa40b62fe53614b87476bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5643077
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>