[RPP] Only redraw the parts of label overlay when needed
Previously, the code that renders all parts of the label overlay was in
`afterOverlayUpdate` and was called each time after the overlay was repositioned.
This CL splits the rendering of the parts of the entry overlay into different methods (label, connector and the entry wrapper).

Now we only redraw the entry label if the label has changed, the wrapper box only if the dimensions of the entry have changed and the connector is only drawn in the constructor since it's shape never changes.

Bug: 339802669
Change-Id: I17a7521879660e1abe411c25be9cd4cf3462bdf9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5639956
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>