[RPP] Handle entry labels input and paste
Limit the possible input for an entry label to make sure that the there won't be labels so large that they can significantly increase the file size. In the future, we might also consider limiting/resizing labels from the UI perspective.

Also, do not allow multi-line labels. Now `Enter` key will be used to "lock" the label and put it our of focus (blur).

Remove `set label` function and instead pass the label in the constructor when creating the label since we will always need to pass the label value into the label overlay once (when it is loaded from a file or just created).

Bug: 339802669
Change-Id: I429a6ad5dfc1cfad91a8437b0bf4a840858dc01c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5645895
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>