[RPP] Use synthetic width for selected layout shift overlay
Layout shifts are added to the flamechart with a synthetic width mapped
from 5,000 microseconds. Otherwise, shifts would be barely visible
because they are instant events and have no duration. We need to use
this synthetic width when calculating the width of the overlay
element added when layout shifts are selected (instead of its duration).

Fixed: 349253903
Change-Id: Id1650ce9d4d8381a1feb83684e85595ad20c4300
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5648211
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>