[sdk] Update 'bindings' decoding to mirror latest spec text
The PR https://github.com/tc39/source-map/pull/106 changes the
'bindings' encoding of the "Scopes" proposal to encode the number of
sub-ranges first. This way, we don't have to look ahead in the VLQ
stream to decode bindings.

This CL aligns the DevTools implementation with the PR.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Ice70c87b4176cfbe23a401b2b336fab144ca99ff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5660062
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>