[RPP] Create LCP phases overlay
The overlay is triggered by a click of the insight in the
sidebar.

TimelineFlameChartView.ts constructs the overlay and calls
Overlay.ts which handles how the overlay is defined and positioned.

TimespanBreakdownOverlay.ts largely borrows from TimeRangeOverlay.ts,
with specifics to handle the TimespanBreakdownOverlay's labels etc.

This also fixes a little bug in LargestContentfulPaint.ts.

Bug: 344550388
Change-Id: Id8fd0d31a943495f63b8555bba90ecac340a03f4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5628319
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>