RPP: position tooltips above overlays
This CL fixes a bug where the tooltips were being rendered below the
overlays, and therefore the selected entry highlight was shown over the
tooltip.

To fix this, we now enable the FlameChart to use a custom HTML element
for its tooltip, and we pass it one that is rendered in the DOM above
the overlays. This ensures that the tooltip renders above the overlays.
Without moving it to this level, we couldn't control that via z-index as
the flamechart is within a different stacking context.

Note that this CL does not move the tooltip to be an actual overlay;
that is a larger piece of work that at this point may not be worth it,
this CL allows us to ensure the visibility of tooltips on top of
overlays. We can migrate tooltips to be an actual overlay in the future
if we need to.

Fixed: 349802751
Change-Id: Id10694cf6bedb3d1fc28ad46484b70dce3c27a2e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5676725
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>