[ve] Fix logging of the clear filter button.
The clear button is actually an uncle of the text field, but we'd like to log it as a child.
This doesn't work with parent provider out of the box, as we scan DOM tree in BFS order, and the text field is not yet processed when we handle clear button. Adjust the logic in getOrCreateState to handle changing parents in case of parent provider.

Bug: 348173254
Change-Id: I9601da151eff059804b24c68c780518f64cebea9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5677284
Reviewed-by: Simon Zünd <szuend@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>