[RPP] Annotations in ModificationsManager
We will need the list of user created Annotations available in the Overlays, Sidebar and ModificationsManager.
To achieve that, create the annotations Map in the ModificationsManager and let ModificationsManager update the other components.

With this change, creation and deletion of the user created annotations should happen through ModificationsManager.

Bug: 349530077
Change-Id: I7bdc65f63ad31621ece8299b443b5ff1cf248543
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5680946
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>