RPP: add navigationId to layout shift clusters
This will more easily enable us to group layout shift information in the
UI when a trace has multiple navigations. You could do this via
timestamps and grouping, but we can easily do this in the handler to
avoid anyone else having to do that work.

Bug: 348591291
Change-Id: I2beb0928c37664a84f6d130960fcf187e60c0062
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5683287
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>