Always initialise traceBounds in test
This tidies up some of the console noise when running the unit tests.

Historically we supported passing `false` here, but there's no need to
actually not-init the bounds manager, and the test modified here didn't
do it for any particular reason, so we might as well always do it, which
is good because it more closely aligns the tests with what happens in
reality when people use the perf panel.

Bug: none
Change-Id: Ieee39770c54c476f05628be4feb731d5f363b694
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5687957
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>