RPP: cache SyntheticEventsManager on rawEvents array to fix tests
We have had some tests on CI that fail when run as part of the overall
DevTools suite, but pass in isolation.

One of the reasons this happens is because of how the
SyntheticEventsManager is initialised. In tests, because we cache and
don't re-run model.parse() on the same trace, we end up not having the
right active SyntheticEventsManager for a given trace. The test will
then pass if run in isolation because when it is the only test running
it has the correct data setup.

To fix this I modified the SyntheticEventsManager; now it caches
instances based on the rawEvents() array, and tracks the active one. We
did support fetching by an index, but the API didn't actually ever
return the user what the index for a given trace was, so this was
useless, and we only ever care about the active one anyway.

Bug: none
Change-Id: Ie709b8bff370c395e3eeed2029fddd5086a3b2b4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5687953
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>