[RPP] Use TraceParsedData instead of partial type
The majority of usages of the TraceProcessor are from running all
available handlers, therefore using a generic type to specify what
handlers are run is now mostly overhead and gets in the way of
comprehension. Now that the trace engine migration is finished this
can be loosened to being all the handlers.

Remaining is the use case in NetworkPanel which creates its own
TraceProcessor with only a single handler. The generic types remain
for this usage, but callers must now enforce the type of traceParsedData
manually.

Bug: 313905799
Change-Id: I15d5cc03449d1e5e050197099c5fd39c7ee4ae3d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5691518
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>