RPP: Remove outside-tracks hover handling
This handling was added to support the edit icon, but as that's removed, this has only been triggering unnecessary redraws of the flamechart (when the mouse is outside of the tracks.)

Functionally this changes nothing. It's a small perf win, and boon to debugging.

Bug:311439339
Change-Id: I0db4046b8e00dca3241d0ce41fef7fd71e7b0736
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5696885
Auto-Submit: Paul Irish <paulirish@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>