[scopes] Add new ScopeChainModel class
This CL is the first in a series that attempts to disentangle the
different responsibilities around the scope view:

1. Each call frame for a debugger pause can have a number of different
scope chains depending on which debug info is applied. The debug info
can arrive before or after we are paused and can also take some time to
process. Furthermore, source maps can be enabled/disabled dynamically,
both cause different scope chains to be available.

2. The UI has a single call frame in which it is interested. This is
the call frame the user selects and defaults to the top-most upon
a pause. The UI is only interested in the scope chain of the currently
selected call frame. This concept of "currently selected callframe"
shouldn't leak into the models layer as it is strictly a UI thing.

To this end we extract a new class from ScopeChainSidebarPane.ts:

A ScopeChainModel on the models layer. It is reponsible
to calculate and update the scope chain for a given call frame
based on the status and availability of debug info.
It notifies interested parties whenever the scope info changes
for the given call frame.

This Cl adds the ScopeChainModel but currently only sends updates
that "something" changed. The code to actually calculate the
scope chain will remain in the sidebar for a bit longer, until
we have the necessary interfaces/infra in place to move it
into ScopeChainModel.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Ic14f1641ff7a5724f47aff0de0513a7a8784c567
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5700945
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>