[sources] Use ScopeChainModel in the ScopesChainSidebarPane
This CL builds on top of https://crrev.com/c/5700945 and utilizes
the new ScopeChainModel class in the sidebar pane to get notified
when the scope chain for the currently selected call frame gets
updated.

We do this by listening to the CallFrame flavor update and creating
a new ScopeChainModel for the currently selected call frame.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I84941a8d007257f02f023977fdf7df92269b7978
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5701560
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>