[scopes] Move resolving thisObject and scopeChain to ScopeChainModel
This is a first step in moving the scope resolving business logic
from the view layer to the model layer.

The CL moves the two 'reslove*' calls to the ScopeChainModel and
the result is sent along the 'ScopeChainUpdated' event.

The CL also renames the ScopeChainSidebarPane.update method to
'buildScopeTreeOutline', as that is what it's doing now.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I5e9a3473583058b21e9a7a63c3048df8301e95ab
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5707448
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>