RPP: tidy up loadingComplete and setModel
This CL attempts to clarify some of the behaviour in loadingComplete and
setModel within TimelinePanel.

Because setModel is called everytime a trace changes, most of our code
relating to the import of a trace should run here, so that it runs when
the user uses the history dropdown to change traces.

There is more to be done here - I don't like the repeated checks for
`traceParseData` within `setModel` - but I don't want to introduce loads
of changes all in one CL.

Bug: 352750032
Change-Id: Ia062f1ddd75804f1f58beba458e2b5f4b9092aea
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5713249
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>