[ve] Fix memory leak in the Non-DOM state
The leak happens because we register non-DOM child loggables for some DOM nodes that never appear in the view port. Previous implementation used to hold strong references to both the child and the parent, which was released when child impression was logged. Since this is not guaranteed to happen, we might accumulate lots of such entries.

Current code stores non-DOM loggables in a weak map keyed by parent, so that they all could be released as soon as parent is released.

Bug: 352496178
Change-Id: If32ff440b95226a4c6eec5742917e1773844654c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5713929
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>