[Animations] Make sure that animations that were supposed to be in the same group are in the same group
Preivously, when the `animationStarted` events come
in different times, even though the startTime (or the
timeline & axis for SDAs) for such animations were the
same, we weren't adding them to the same group.

This CL changes that logic to append newly coming
animations to already existing animation groups
if they satisfy the condition for being in that group.

For this, we also got rid of using `animationCreated` event.
Previously, it was used to make sure that if an animation
is created, we waited until it is started to create animation groups. However, this is super prone to errors & future
breakages as now that there are animations that might
be created but not started for SDAs (See this example[1]).

Drive by:
* Replay a preview group if its animation is updated.

Bug: 349554230
Change-Id: I4266133c859a94c914ed2fba5bcabfad7ad1502a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5718130
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>