Add platform-X to `html` not `body`
When trying to use the new `--sys-typescale-X` variables in
designTokens.css, I discovered that they did not have access to the
`--default-font-family` variable.

This is because that variable is defined via the `.platform-X` classes,
which were being added to the `body`. However, the design tokens are all
added to `:root`, which is the `html` element, and therefore are higher
in the tree than the `--default-font-family` variable.

This CL moves the adding of the `platform-X` class onto the `html`
element, which means the `--default-font-family` can be accessed from
`:root`.

It does this by introducing a new method. At first I moved the class and
the appending of all the common stylesheets onto the HTML element, but
this breaks specificity across DevTools because the common sheets are
applied to the HTML element rather than the Body and this breaks some
UI.

So instead, a new method is added so that we still append stylesheets to
the body, but add the class to the HTML element. This is all that is
required, because `:root` in CSS sheets like `inspectorCommon.css`
refers to the HTML element, and so does any custom styling in
`platform-X`, such as setting the font.

Bug: 355622977
Change-Id: I4a4269534dc86b3db2b5dac32cd15d0a8f4897fb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5741479
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>