RPP: add ENTRY_OUTLINE overlay
This CL adds an overlay for showing borders around entries, most likely
to highlight problems (e.g. render blocking requests). That's the only
considered "reason" for one of these now, but I have coded it in as a
property as I could see us having multiple in the future.

There is now some overlap between this + the entry label overlay...I'm
wondering if we use this for the outline and then use the EntryLabel
overlay for the connecting line and the label itself...but unsure and
will figure out in a follow-up, and see what alinavarkki@ thinks...

Bug: 343388196
Change-Id: Icecc3fc62cfd3537439bf959243ebb9447c961c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5745075
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>