[Settings] Add Switch component for the privacy framework settings UI
Screenshots: https://imgur.com/a/8H8mDCG
See the bug for more details.

For using this component for the first time, we'll need to add the files to `devtools_grd_files`. I didn't add them since we don't use this component yet and there is no need to include it in the bundle.

Fixed: 355415645
Change-Id: I48454b9031fa8a9ca6e727303d524f613f5c9907
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5745073
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>