[RPP] Fix the hardcoded 0 in cached network requests
Before, when we see a network request is cached, we will hardcode a few
timings to 0. But this is not really correct, for example the
|Queuing and waiting| time are not really 0.
So the fix can be if the |timing| field exists, we use the information there,
otherwise fall back to 0.
Also for some legacy traces, the |timing| field is incorrect if it is from
memory cache.  So for those events, we will ignore their |timing| field.

Screenshot of before & after: https://screenshot.googleplex.com/6Vzo45WrVUvGmtv

Bug: 355392317
Change-Id: I4191c7e7c484029d26750a992e7ab6c197f67794
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5756761
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>