[RPP Observations] Add basic info nuggets to metric cards
https://screenshot.googleplex.com/4Ctja2pv2iAsRWG

The code for the info nugget strings is auto-generated by a script. In
theory, we shouldn't need to run this script very often so not worth the
effort to add it to the formal build process IMO.

This required a pretty large refactor of the `MetricCard` component,
but it's a worthwhile cleanup.

Bug: 313906438
Change-Id: Ia6c7f84954ba71e358031f84f48eaaad22d69ee1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5757940
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>