RPP: create base class for sidebar insight
The two insights we have built so far and the many more we will build
had a lot of duplication in them - mostly plumbing to pass through the
active insight, navigation id, trace data, and insights. This CL creates
a base class that the insights will extend. They can still define their
own methods and ultimately define the render method, so they maintain a
lot of flexibility, but it means we don't have to duplicate the setters
for all the common data.

Once this lands, I will follow it up with a CL to add documentation to
the folder to document how to add a new insight and what is provided by
the `BaseInsight` class.

Bug: 351758501
Change-Id: I4902ec20da3dd6203bbdb84954b23e8639fddf61
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5756850
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>