RPP: store user consent differently in OTR profiles
The consent for CrUX needs to persist differently if the user is in an
"OffTheRecord" (OTR) profile.

1. In OTR, the setting needs to not persist, and so we use
   "sessionStorage".
2. In "normal", the setting needs to persit, so we use the DevTools
   "global" storage, which is the default for settings and syncs values
   to the profile.

This CL uses the `isOffTheRecord` field which was added to the host
config in [1]. If the config is missing, or `isOffTheRecord` is true, we
use the session storage. Otherwise, we use the default.

I wasn't sure if we should instead use two settings for this rather than
toggle the storage type, but because the internal registry of settings
is create from scratch every time, it's OK to have the same name with
different setting storage locations. The only thing DevTools doesn't
support is changing the storage location in-flight, but we don't do that
because it's always a different window if the user then moves into
incognito.

1: https://crrev.com/c/5755942

Bug: 356612551
Change-Id: I23ee806f3e7bb36dbb087607669afcb64cc1b69b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5756492
Reviewed-by: Adam Raine <asraine@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>