[RPP] Remove SyntheticTraceEntry
This is an interim change created for making an otherwise complex change
simple.

Events typed as such are not synthetic in the first place. The type was
used to mark events handled by the RendererHandler, but there isn't much
benefit in doing so since in practice any arbitrary type of event could
be handled by it.

The one utility this type had was allowing events to have a selfTime.
However, we don't want to add this property to raw events. Instead,
because this property exists if and only if the event is part of a tree
(nestable / sync like), it makes sense to store inside the Node object
created for events in trees. This step will be done in a follow-up, in
the meantime I moved the selfTime property to the ubiquitous
TraceEventData, (only temporarily, to keep CLs simple).

Bug: 330204386
Change-Id: I4fc6b15575cfc79b477ce76f60f82c2db361164f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5756955
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>