RPP: limit max invalidations stored per event
We have seen instances where traces with a large amount of invalidations
can cause the trace engine to take a long time processing and hold onto
a lot of memory - even causing LH to OOM in Node (https://github.com/GoogleChrome/lighthouse/issues/16111)

In these instances we've seen traces where one Layout event (as an
example) can be linked to thousands of invalidations. Once you get to
this point we aren't helping the user to debug by showing them a list of
every single invalidation. Therefore, to fix the performance of this
case, and to make the UI not a huge list, we now cap the number of
invalidations we will track. We also count the total, and in the UI we
now show this total.

By making this a configurable value, we can also allow a user to disable
Invalidations by setting it to 0. In this case we shortcut the handler
and will not process any events.

Bug: 352680159
Change-Id: I63c8dc41577e4a350788885a7bc23967ace0c10c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5766691
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Adam Raine <asraine@chromium.org>