[RPP]: enable overlay for LCP discovery insight
This constructs the overlay using 1) the candy stripe bar, 2) the error
entry outline, and 3) a singular timespan breakdown. The singular
timespan breakdown is built based on the presence of its entry.

This also fixes a slight bug where the candy stripe bar would follow
when scrolling. When the overlay's event would be hidden, it would
cause the function to end early without correctly updating the style
of the bar.

some things to note for follow-ups:
- need to ensure zooming takes into account vertical positioning
- scrolling down should squeeze the timespan box instead of hiding

Bug:351758501
Change-Id: Id2704a85487d6c87b59edf507fba72ed94a76b29
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5767596
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>