[RPP] Entries Link arrow snaps to a hovered entry
Pass the hovered entry to the Event dispatched from the Flamechart to create and update the annotation and update both annotation and overlay with the hovered entry as the second entry in the link.
When the second entry exists in the annotation, snap to the edge of it instead of following the mouse.

Bug: 339802345
Change-Id: I21ad8364adafb835297034a4c5825045d9550018
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5783578
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>