Heap snapshots: store class names in unused bits of detachedness field
A proposed upcoming change will introduce a new case where the name of
a heap snapshot node doesn't match its class name, and the class name
can't be easily computed from the node name. For that change, we'll need
a way to store the class name separately from the node name.

With this change, DevTools will store the index of each node's class
name in the previously unused bits of the detachedness field. However,
the detachedness field is optional: there are tests in Blink that do not
include that field, and there may be snapshots produced by external
tools without that field. To work around that problem, this change
implements storing the detachedness-and-class-index value either in the
nodes array or in a separate array allocated for that purpose. I think
that this small piece of added complexity is pretty well encapsulated
and worthwhile for the memory savings in the common case.

This change is intended to not affect observable behavior, but it's
worth mentioning an internal implementation difference: class names
based on node types, such as "RegExp" and "(system)", are now included
in the string table rather than represented as negative values in the
classIndex() function.

Bug: 356183770
Change-Id: I75248ca7b73250621120f71de48c9df2913d9b6e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5782661
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>