RPP: update visual logging for observations
This CL updates the logging to log clicks on the various buttons used to
configure / opt-in / opt-out of field data.

It also adds an attribute to the metric elements to show what they are,
so we can see if we rendered data.

Bug: 344852012
Change-Id: Ica6e8b16469f0629f4085aec29a940e062d01c3b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5788585
Reviewed-by: Adam Raine <asraine@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>