Memory: Move the primary page target to the top and select it by default
Though the code in the `isolateAdded()` method of `IsolateSelector`
intends to add the primary page target ("Main" target) at the top of the
list of JavaScript VM instances and select it by default, this is not
actually happening because of the following:
1. The check for the target being equal to the root target is false even
for the main page. The root target has type 'Tab' while the main target
has type 'Frame'.
2. The `isMainThread()` method returns false even for the main target
because the target ID is a hex number which will never match 'main'.

This CL fixes the above by comparing the target with `primaryPageTarget`
instead of `rootTarget`, and updating the condition for selection to be
based on the index being 0, instead of another redundant check.

Also replaces the other use of `rootTarget` (in `updateTitle()`) in this
file since that condition will also never be satisfied. The effect of
this change is that the suffix ": Main" is not added to the main target,
and instead only the URL of the page is shown, as originally intended by
this code.

Verified that the main page is moved to the top and selected by default
for pages with and without shared and service workers.

Fixed: 40669896
Change-Id: I59b3d97bf8ab53f0907715884779df14deab7c11
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5790116
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Auto-Submit: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Reviewed-by: Seth Brenith <seth.brenith@microsoft.com>