Memory: Sort children when populated and not just when expanded
Currently, when we change the sorting by clicking on a column header,
the child elements are recursively sorted only when they are expanded.
When a node is expanded for the first time, its children are added in
the currently set sorting order and the `populated` field is set to
true. After this, if the node is collapsed and the sorting is changed,
the recursive sorting does not apply to the child elements. When the
node is then expanded, the children do not get repopulated and hence
they show up in the earlier sorting order, causing the linked bug.

This CL fixes it by changing the condition for sorting the children from
expanded to populated.

Fixed: 361002653
Change-Id: I563648e67687169bc457b6fc8505dd65027ba4ba
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5804462
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Seth Brenith <seth.brenith@microsoft.com>
Auto-Submit: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Commit-Queue: Simon Zünd <szuend@chromium.org>