[RPP] Add render blocking sidebar insight
For now this shows FCP estimated savings, this could be changed later
when there is a clearer direction to take. The insight has both LCP and
FCP, but there are times when LCP savings are 0.. Additionally we're
still exploring how we are approaching FCP in rpp.

https://screenshot.googleplex.com/9n8eVu2H2TjdwYy

Bug:328489888
Change-Id: Id98ffc49c367b34bc85e74c1d739330ee5a2f787
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5791753
Commit-Queue: Adriana Ixba <aixba@chromium.org>
Reviewed-by: Adam Raine <asraine@chromium.org>