[RPP] Enable creating link between FlameCharts
Enable creating Annotation Link between Main Flamechart and Network Flamechart. To achieve that, move the logic that updates the link in the process of creation to the TimelineFlameChartView.
Previously, Flamechart contained the current link state, but we have mutiple Flamechart so it wasn't possible to create a link between them because only the Flamechart that created the link was aware of it. Now Flamechart is only responsible for creating the link and TimelineFlamehartView handles updating and deleting it.

Also, pass elements for both Flamecharts to OverlaysImpl to get mousemove coordinates from them

Video: http://screencast/cast/NjMyMDU2MjM2NzM2NTEyMHxmYzZhZTY2Mi1jMQ


Bug: 339802345
Change-Id: I8c4753c85aaad8e5210e6498700bbc47eb98c04c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5803962
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>