[RPP] Display the Link annotation in the Sidebar and show annotated entries in the Sidebar in their FlameChart colors
To enable to easier identify the annotation, display annotated entries with the same background as they appear in the Flamechart.
To achieve this, build a map of entries -> colors that is passed to the Sidebar together with the list of annotations.

Also add the Entries Link annotation to the Sidebar.

Video: http://screencast/cast/NTY5MjQwNTk1MjAyMDQ4MHwyYzUzNzJmZS02Zg

Bug: 339802345
Change-Id: I0e523032262793b70c8d4d22acf20ef58594a28c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5803647
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>