[SettingCheckbox] Enable text content override
The setting title does not always provide enough context on its own. In
this case the setting for the network cache does not specify what kind
of cache it is referring to. It is useful to control the text label of
the checkbox for this purpose.

Bug: None
Change-Id: Ic072811c40a23686b26082b100c9e95501ba53ff
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5807954
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>