[Freestyler] Merge thought and action
As per the design these appear on the same block so we need to change
our logic a bit. We use the same id to signify that the step need to be
the same and we pass though any data from the though to the action.

Note it may be possible to have a Action without a Though.

Bug: 358324719
Change-Id: I7d8940d3414da4357dc151f503e1f69a4894ab28
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5804324
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>