[Freestyler] Split the logic of how we show messages
This decouples the structure coming from the Agent with what is
given to the View, that way we can keep the needed properties
directly on the object that uses them.

With this we now consider Step only intermediate part, this means
that an answer is it's separate entity.


Bug: 358324719
Change-Id: I0b0e36d9f93fc68c1649fe759630be3cb618d714
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5803651
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>