[Console insights] Fix for backwards-compatiblity for remote debugging
This is a follow-up to https://crrev.com/c/5782756 and
https://crrev.com/c/5753004.

When remote debugging an older version of Chrome, the host config's
shape needs to be transformed to its old shape. The old shape
contained both an `enabled` and a `blockedByFeatureFlag` attribute,
which both contained the same info. Since older versions of DevTools
rely on the `blockedByFeatureFlag` attribute, it is being added in this
CL.

Bug: b/348136212
Change-Id: I5c32ff952643bb00ed252b8f1b0efe9772da7188
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5816533
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>