[RPP] Add navigation to synthetic layout shift
We calculate the navigation id for the cluster, so we can add this to
the Synthetic Layout shift if we have it. It is useful knowing the
navigation when we are displaying Layout shift insight data within the
panel. For example the new Layout shift details summary panel contains
insight data.

Also a drive by fix for calculating the worst layout shift to use score
instead of cumulative score.

Bug:329735547
Change-Id: Ice9e3724d7f0fbb234c11d9c2e91fdb4631a69c9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5818058
Commit-Queue: Adriana Ixba <aixba@chromium.org>
Reviewed-by: Adam Raine <asraine@chromium.org>