Remove SettingTracker
The SettingTracker was needed for the opt-in vs. opt-out for console
insights. Opt-in does not exist anymore, the `console-insights-
enabled` setting is true by default, which means the feature is
enabled by default.
The ability to configure this via Finch or command line has already
been removed, the SettingTracker is already unused and can be removed.

Bug: b/350668580
Change-Id: I40ba66b7df6bf5da8d885204add3dacdae515daf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5822712
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>