[sdk] Add scaffolding for SourceMapScopeChainEntry
This CL adds a new implementation of the `ScopeChainEntry` that we'll
use to provide "Scopes" proposal scopes to the UI.

The impl is not really interesting yet as it only handles the l10n
for the spec pre-defined scope kinds.

Follow-ups will add a RemoteObject implementation that'll utilize
the binding expressions, and functionality to construct a full
scope view just from the SourceMapScopesInfo.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Ia9f7faf7ca07481718ba8ee450d8929e071d73d0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5828826
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>