RPP: allow annotation labels on frames
This CL allows annotation labels to be added to frames (in a follow-up
we will also add support for frames when drawing connections).

This CL turned out to be slightly more complex than anticipated; the
LegacyTimelineFrame type (named as such because we expect to move frames
to being based on AnimationFrame events soon) was not defined in
`Types.TraceEvents` but instead within the `FramesHandler` (mainly
because the FramesHandler is mostly based off the legacy TimelineModel
implementation). However, we needed to pull out a type into
`types/TraceEvents` so it could be referenced in the right place.

So, I did that, and then after that it was a case of adding
serialization support for these frames so that annotations attached to
them can be stored in a trace file and restored when it is imported. I
then updated the overlay code to be able to handle being attached to a
frame.

Fixed: 362950098
Change-Id: I4b345719fcf8293ea3302c4e1c675d51cf4e4212
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5828601
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>