[Freestyler] Make changes unique
With this change, a unique ID is generated for each element that
is changed. All changes to the same element are merged into one rule.
Only the initial selector for the element is recorded. If the selector
changes later on, it will not be reflected in the change.

Bug: 361742726
Change-Id: If30427041c9c2e629c7bf7be780bb413a4047da0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5831210
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>