[cleanup] Remove "Highlight important DOM properties in the Object Properties viewer" experiment.
This feature never made it past the experiment stage and while we believe
that it's beneficial to highlight important DOM properties, we aren't
convinced that this approach is the right one. Since we have no capacity
or intention to work on this experiment, we are removing the code now.

Doc: http://go/pinned-properties
Bug: 40267030, 40840064
Change-Id: Ib2d0f287834ca9ee87d957bd64f77a450c322f46
Fixed: 315126594
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5836838
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Victor Porof <victorporof@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>