[scopes] Enable the 'Scopes' proposal backed scopes view
This CL enables building the scope view completely from the scopes
info provided by the source map.

This feature is still behind an experiment, as we only parse the
scopes info if the experiment is enabled in the first place.

The actual functionality (and tests) were implemented with
https://crrev.com/c/5833561 and https://crrev.com/c/5829405.

R=kimanh@chromium.org

Bug: 40277685
Change-Id: Ie7ea2a1ae679df780dea7464adfb9d9b9f3976fc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5836933
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>