[sdk] The SourceMapManager shouldn't refresh source maps itself.
The responsibility of the SourceMapManager should be to purely react to
`attachSourceMap` and `detachSourceMap` calls from the debugger and CSS
models. These models already take care of reacting to main/primary frame
changes and update the `SourceMapManager` appropriately.

This fixes a race between the `SourceMapManager` reacting to
`InspectedURLChanged` and the `CSSModel` reacting to
`PrimaryPagechanged` events, which causes test failures with the
RenderDocument experiment.

Bug: 357977710
Change-Id: I666d265acf514eb7bc6f679a18b07bd78a64e395
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5839523
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>