[RPP] Checkbox to hide annotations
Add a checkbox that allows to hide annotations to the Annotations tab in the Sidebar.

If the annotations are hidden and 'create new annotation' was called either through the context menu or a shortcut, make the annotations visible.
Since the time range selection could be used just to check the length of the selection, make annotations visible only when a label is typed for the selected time range, instead of when it is created.

Video: http://screencast/cast/NjA3NjE5OTkzNDYyMzc0NHxlOGYwMDMyMy1iYw


Bug: 362676646
Change-Id: I93f887b3e92826a14f41fea9eaa2bb6e560a6f0c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5823358
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>