[RPP] Construct tracingCompletePromise well before it is needed
It seems feasible that 'tracingComplete' could be invoked before
'stopRecording', which would cause 'waitForTracingToStop' to never
resolve. I was not able to reproduce the issue on purpose, so hard
to verify as a fix.

This refactor creates the promise just before starting the trace, so
now it is obviously always in a valid state.

Bug: 364921460
Change-Id: Ie293eeaf44d1abfe43fe9c870edecbc5bef21f2a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5840396
Commit-Queue: Paul Irish <paulirish@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>