RPP: start on general EntryName solution
We have a lot of code in different places for naming entries and what
their user visible title is. This is the beginning of a component to
unify that and have all places in the UI use the same implementation to
ensure we are consistent across the entire codebase.

Bug: 365047728
Change-Id: Id8c9c8acb076eb00d864f202a1e79b5c95e50bb9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5838373
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>