RPP: fix loading of events and eventByKey lookup
I discovered an issue where we were not correctly restoring hidden
entries which I tracked down to a binary search issue in the
EventsSerializer. We tried to binary search the events to be more
efficient because the arrays can be long. Howeveer it seems that the
events are not ordered in ASC order (at least, not if you combine the
sample index and the node id). Therefore this CL moves us back to a
`find` call, and updates the predicate to check for strict equality
(which also fixes a bug where we found find the wrong ProfileCall event
for a key).

As a drive-by improvement, I also added some error handling to the code
that applies modifications when a trace is loaded, ensuring that if we
get dodgy keys in the trace file, we just ignore them and move on rather
than crash the panel.

This CL also moves the call to update the active ModificationsManager to
earlier in the lifecycle, previously it was being done after the flame
chart was drawn so they were not actually applied. Now it is done before
so when you import a trace which has valid hidden entries, they actually
get hidden.

Bug: none
Change-Id: Ib2e2b859e4f4ac07fcd10b1e48f8c76c66c02687
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5842406
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>