Move "annotations hidden" setting into timeline-meta
This broke the worker entry point because this setting was defined in
the elements meta which is not loaded in the worker app. Instead let's
define this setting in the timeline meta file so it's always loaded when
the performance panel is included.

Fixed: 361515458
Change-Id: I6b9637c498467b0e380629262adaa5a50366355a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5845028
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>