[AI Settings] Show toggle in "off" state when disabled
If a toggle is disabled (= user cannot switch its state), the
corresponding feature is not available to the user. The toggle should
therefore be in the "off" position, even if the underlying setting is
enabled (and only "disabled by condition").

Bug: 365706617
Change-Id: Iaa9ed1b44700f20aa96aacfbc4ec4872e504158d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5853476
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>