RPP: standardise more entryTitle code
This is another CL in the road to unifying all our logic for deciding
entry display names; we have a lot of code for this and as a result in
some UI we show different names to in other places.

timeline/components/EntryName.ts is now the home for the method that
determines the user visible name. This CL updates some parts of the UI
to defer to it rather than customise it.

The idea is long term we can get rid of all entryTitle() methods
(especially in all the appenders) and just call into EntryName directly,
but I am doing this work incrementally to avoid huge CLs.

Bug: 365047728
Change-Id: Iec94cb0b3ba46343e5eb7efd0dda5398931458ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5855506
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>