Settings: introduce `getIfNotDisabled`
A setting has a boolean `value` and an independent boolean `disabled`.
This CL introduces a getter which only returns the `value` if the
setting is not disabled.
This makes it explicit at the call site, to take the potential
disabled state into account.
When using the regular getter, one has to be aware of the possibility
of the setting being disabled and to not forget handling this case.

Bug: 365706617
Change-Id: Ia90117b458dc7f7661050f0e136e01c99b7c7d2c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5854599
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>