RPP: entryTitle consolidation
This CL removes a few custom entryTitle methods in favour of using the
canonical function.

It also makes the `highlightedEntryInfo` of the track appender optional;
most of them were returning the title and formatted time so we can just
do that by default and let an appender override if required.

In a follow-up CL I will go through all the other appenders and change
them accordingly.

Bug: 365047728
Change-Id: Ie9b35b7633c12222c0c3c79ff04d71c582a12592
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5860635
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>