Ensure NetworkThrottlingSelector updates active items
This fixes a bug where the NetworkThrottlingSelector would not pick up
on the latest changes to the network throttling if it is not mounted
when the change occurs.

This can happen if the user:

1. Loads up the performance panel
2. Navigates to another panel (network / performance insights)
3. Changes the network throttling setting
4. Returns to the performance panel

Because the listener gets removed when the component is disconnected, it
would not detect it. The fix is to immediately in the connectedCallback
add the listener and call the change handler to detect any changes that
happened whilst the component was not listening.

Fixed: 366886102
Change-Id: Iabe773cb9381466e713fa167b24f93d0b3426457
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5864295
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>