RPP: ensure entry labels render above the selected outline
Otherwise it can become hard for them to be read if the selected outline
goes through the text. In time we might want to do this for more
overlays, but for most of them the selected outline does not obstruct
them as much as it does for labels.

Fixed: 365513648
Change-Id: I2193e92ca36b6f1a876eb789fa05e8adc299cd43
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5872601
Reviewed-by: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>