[Dialog] Don't use animation offset when placing pointer event in modal bounds
When the dialog is shown as a modal, it is not animated when its open.
Thus, the extra offset taken for when the animation takes place must not
be considered when comparing its bounds to pointer events.

Fixed: 366309861
Change-Id: I05f8223cd5c9b4cd0e43c12c768ae3e77bc42db0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5875964
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>