Add support for dynamic static tags in Lit.html
`<${tag}>` is now supported, where `tag` is a `LitHtml.literal`
that varies within the same template string declaration.

Previously `htmlWithStatics` would cache the result of flattening
the template strings array, which assumed that the static values
within would never vary. This disallowed dynamically selecting
which component to render, and required an explicit switch/if-else
instead. Now, the values of the static literals are used as part of
the cache key.

This CL also includes an RPP refactor to simplify how we are
rendering the insight components, and an additional drive-by to use
a `data` setter for these components.

Bug: 368135130, 40181249
Change-Id: I8f6332d6adeb6fc7a447df18271d5456bd7f066e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5874566
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>