Pick debug symbol to use when multiple are available for WebAssembly
The v8 back-end now informs us about all available debug symbols
that exist for a WebAssembly module. In this case, the front-end
may now actively choose the debug symbol to use, and issues a warning
to the developer that multiple symbols exist, and which symbol it
ends up using.

Bug: 40879198, 369515221
Change-Id: Ib5b73db6b34481823f42dfe697ad6ceab10a7605
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5878368
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>