[Freestyler] Add opted-out state
If Freestyler is disabled in settings, show an empty state view in AI
assistance, instructing the user to enable the feature in settings if
they want to use it.

Entrypoints to Freestyler are shown, even when the feature itself is
disabled.

Screenshot: https://i.imgur.com/4YeALGs.png

Bug: 353441061
Change-Id: I5c4459b45441e3a17b563bb76ef7835bfaba61fc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5882695
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>