RPP: improve selecting event in group via keyboard
This CL improves the behaviour of a user tabbing into a track to select
it, and then hitting enter to select an individual event.

Now we will select an event that is at least 1ms (if possible)
(otherwise it can select a TINY event that is hard to see). We also tell
the user about it via aria-live, and tell them to press enter again to
focus that event.

Fixed: 369976524
Change-Id: Ic8cb3fcc3f7da1e927fdbe3b55215062a8c12617
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5895396
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>