Ignore list setting: Update and clarify wording
The word 'pattern' is used, but arguably incorrectly. Generally
pattern is a glob pattern, where the * plays an important role.
These exclusion rules are not parsed as glob patterns but as regex.

Also added notes to clarify Perf panel impact, and that they're
matched across the full URL, not just the filename.

https://screenshot.googleplex.com/9k4DAeVN9mya98E

Bug: 368222753
Change-Id: Ic8a47e46636a97deb0ee079941b77ee97340c854
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5874285
Commit-Queue: Paul Irish <paulirish@chromium.org>
Reviewed-by: Sofia Emelianova <semel@chromium.org>