RPP: change behaviour of clicking on metrics
This CL updates the metric click behaviour to reveal the relevant entry,
but removes the association to an insight with each metric.

This CL also fixes a bug where we couldn't reveal marker events due to
their NaN duration breaking the logic of the flamechart's reveal.

There is also a drive by to rename `isSelection` to
`isTraceEventSelection` in `TimelineSelection` to make that clearer.

Fixed: 368678568
Change-Id: Id74e9c2ab38f99f4fa28451884cbf32217ebbc00
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5895517
Reviewed-by: Paul Irish <paulirish@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>