[RPP] Do not shift track into the view when clicking on an entry when the connection creation is in progress
When the connection is in progress and an entry from a different track is clicked, the track gets selected, which results in clicking a different entry from intended because the track is shifted. To avoid this, do not shift the track when a connection is in progress.

Before: http://screencast/cast/NjM3ODQ4MDIwMjY3ODI3MnwzOTg5YWUxZC0wYw

After: http://screencast/cast/NjYyMjg5ODUzNzEwMzM2MHw5NDFlM2Y1NS0yYQ

Also do not select second entry if the Connection is not started yet.

Bug: 368245209
Change-Id: I8a83dac9588174e7f8ec03c5d62b90ad593c3ccf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5898916
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>