RPP: fix errors in worker_main entrypoint
This CL fixes an error when trying to use the Performance Panel to
inspect a worker via the worker_main.ts Devtools entrypoint. It was
trying to load the EmulationModel which is not included and therefore
does not register its settings.

There is probably a reasonable discussion to be had about not showing
the live metrics view in this entrypoint, but for now we can easily wrap
the calls to load the model in a try/catch to avoid erroring and
breaking the entire panel.

Fixed: 361515458
Change-Id: I0cf6691bcae9f7ab73714273c78a074b1bcfcaeb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5837022
Reviewed-by: Adam Raine <asraine@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
(cherry picked from commit 02fca728b2a8e43897b708811d57aca1d2130970)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5899733
Commit-Queue: Adam Raine <asraine@chromium.org>