Insights: pick INP from the set of non-nested events
We had this bug which said that overlays did not appear for
interactions; in actual case the problem was that the Interactions
insight was not picking from the same set of interactions. In the panel
we de-nest events and pick from that set (by definition once we have
de-nested, the ones we are left with are the longest). But in some
traces you can have two interactions which have the same duration, and
then I think depending on how you sort + pick them, you can end up
getting one that wasn't visually rendered.

The fix here is to ensure the insights uses the de-nested set to find
the longest interaction.

Fixed: 366040373
Change-Id: Id739c88222d1ce6c401c5de6ce42271e562e5a5f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5904620
Commit-Queue: Paul Irish <paulirish@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>