RPP: fix bugs with timespan breakdown overlays
This CL started out as a bug to fix the visual display of timespan
breakdowns with an odd number of phases (e.g. INP breakdown) where the
wrong sections were visually rendered higher.

This CL fixes that, but also a bug I found where the component would not
update and re-render correctly when you interact with it via the Insight
Table component in the UI. The problem was that hovering on the
component would update the overlay, but the code to set the width of the
overlay sections would run before the overlay had been given the up to
date sections. Therefore the widths would not be set correctly and the
overlay sections would have incorrect widths.

To fix this, I updated the update() process for overlays. It's now done
in 3 steps:

1. before positioning hook - this is where you can update an overlay
   with the new data
2. position overlay - this is where the new overlay is positioned
3. after positioning hook - is this where an overlay can run any code it
   needs to run after it's been put into its new place

Now, with this structure, we update the timespan breakdown sections in
(1), so that the position code in (2) runs with the latest data, and
then in (3) we just check the label visibility.

I also introduced a wait for `coordinator.done()` as we use scheduled
rendering within the components. I don't think this will impact the
responsiveness of the overlays because we don't do any overlay
re-rendering when the user is scrolling/panning. If this does, we can
remove this `await` and instead update the components to just render
directly without scheduling, which will be fine as long as we ensure
data has changed before triggering a render.

Bug: 371140062
Change-Id: I3c89cb44919b040b4902e6016509af88f52b044c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5904929
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Adriana Ixba <aixba@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>